requirements.txt
# Core AI
langchain
langchain-community
langchain-core
langchain-ollama
langgraph
chromadb

# Web/UI
flask
pywebview

# Tools & Utilities
duckduckgo-search
mss
Pillow
requests
python-dotenv

# Optional but good to have
colorama